Some refactoring in indexedDB-related scripts; fixes bugs when initializing the db for the first time
Specifically, the bug seems to be that having `console.log(event)` in the indexedDb onsuccess method will
cause the emulator to hang indfinitely unless the page is refreshed first. I'm guessing this has something to
do with how the browser indexedDB implementation syncs to persistant storage, possibly relying on Garbage Collection
somehow, but that's just a hunch.